[STATIC-VECTOR, TYPE TRAITS] static_vector is using a couple of type traits to determine a suitable implementation of rotate(), whether to determine a dlib::hash() overload and a std::hash specialisation. The type traits are general and are therefore moved into a separate header traits.h. A few other type traits dotted around dlib have been moved there too.
